chore: add minimal plugin code
This code allows to use the airdrop plugin in a chain (outside of a
chain will be possible once global plugin PR is merged), by declaring it
in `plugins.yml` file like so:

```yml
plugins:
- path: github.com/ignite/cli-plugin-airdrop@chore/import-plugin-code
```

Once merged the `@chore/import-plugin-code` can be removed. To test
locally, the path can replaced with the absolute path to the cloned repo
directory.

A specific version of ignite is required for now, compiled from the
`feat/network-plugin` branch, because it uses some features I added
for the network plugin.

Also to compile it locally it's probably required to set `GOSUMDB=off` to
avoid the time out issue with sum.golang.org.